Change consult-fd to include matching of file path names
The consult-fd command can match patterns in file paths. However, we
don't want to match the path up to the project dir, just the relative
path starting at the project folder.
For this to be done, we need to include the project's full path at the
beginning of the search expression.

We also need to force a case insensitive search, as smartcase will
pick the wrong case if there are any capital letters in the project's
base path.
This is a trade-off we need to have, as consult-fd lacks
the ability to search in relative file paths (see here
https://github.com/sharkdp/fd/issues/609).